backend/requirements.txt
asgiref==3.4.1
astroid==2.8.5
attrs==21.4.0
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.21.2
gunicorn==20.1.0
inflection==0.5.1
isort==5.10.1
jsonschema==4.4.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg==3.0.3
psycopg2-binary==2.9.2
PyJWT==2.4.0
pylint==2.11.1
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
whitenoise==6.0.0
wrapt==1.13.3
